renaming app.py to ui.py and gutting get_version into its own file
