+++ main.py
@@ -1,9 +1,8 @@
-import pkg_resources
-
 import argparse
 import logging
+from app.get_version import get_version
 
-from app.app import App
+from app.ui import App
 
 logging.basicConfig(
     level=logging.INFO,
@@ -13,19 +12,6 @@ logging.basicConfig(
 
 # Set logging level for botocore to ERROR to suppress stack traces
 logging.getLogger("botocore").setLevel(logging.ERROR)
-
-
-def get_version():
-    # Replace 'your_package_name' with the actual name of your package as specified in the 'pyproject.toml' file.
-    package_name = "aws-stuff-doer"
-
-    try:
-        # Get the distribution based on the package name and return its version.
-        return pkg_resources.get_distribution(package_name).version
-    except pkg_resources.DistributionNotFound:
-        # Handle the case where the distribution could not be found
-        return "unknown"
-
 
 def main():
     parser = argparse.ArgumentParser(
